+++ cax/main.py
@@ -10,7 +10,8 @@ from cax import config, qsub
 
 import pax
 
-from cax.tasks import checksum, clear, data_mover, process, filesystem, tsm_mover
+from cax.tasks import checksum, clear, data_mover, process, process_hax, filesystem, tsm_mover
+
 from cax.tasks import corrections
 
 
@@ -38,11 +39,10 @@ def main():
 
     args = parser.parse_args()
 
+    print(args.run, config.get_hostname())
 
     if args.host:
         config.HOST = args.host
-
-    print(args.run, config.get_hostname())
 
     log_level = getattr(logging, args.log.upper())
     if not isinstance(log_level, int):
@@ -87,21 +87,20 @@ def main():
             config.set_json(args.config_file)
 
     tasks = [
-        checksum.AddChecksum(),
-        corrections.AddElectronLifetime(),
-        corrections.AddGains(),
-        corrections.AddSlowControlInformation(),
-        data_mover.CopyPull(),
-        data_mover.CopyPush(),
-        tsm_mover.AddTSMChecksum(),
-        checksum.AddChecksum(),
-        process.ProcessBatchQueue(),
-        checksum.CompareChecksums(),
-        checksum.AddChecksum(),
-        clear.RetryStalledTransfer(),
-        clear.RetryBadChecksumTransfer(),
-        filesystem.SetPermission(),
-        clear.BufferPurger(),
+        corrections.AddElectronLifetime(),  # Add electron lifetime to run, which is just a function of calendar time
+        corrections.AddGains(), #  Adds gains to a run, where this is computed using slow control information
+        #corrections.AddSlowControlInformation(),  
+        data_mover.CopyPull(), # Download data through e.g. scp to this location
+        data_mover.CopyPush(),  # Upload data through e.g. scp or gridftp to this location where cax running
+        #tsm_mover.AddTSMChecksum(), # Add forgotten Checksum for runDB for TSM client. 
+	checksum.CompareChecksums(),  # See if local data corrupted
+        checksum.AddChecksum(),  # Add checksum for data here so can know if corruption (useful for knowing when many good copies!)
+        clear.RetryStalledTransfer(),  # If data transferring e.g. 48 hours, probably cax crashed so delete then retry
+        clear.RetryBadChecksumTransfer(),  # If bad checksum for local data and can fetch from somewhere else, delete our copy
+        filesystem.SetPermission(),  # Set any permissions (primarily for Tegner) for new data to make sure analysts can access
+        clear.BufferPurger(),  # Clear old data at some locations as specified in cax.json
+        process.ProcessBatchQueue(),  # Process the data with pax
+        process_hax.ProcessBatchQueueHax()  # Process the data with hax
     ]
 
     # Raises exception if unknown host
@@ -147,6 +146,10 @@ def massive():
     parser.add_argument('--config', action='store', type=str,
                         dest='config_file',
                         help="Load a custom .json config file into cax")
+    parser.add_argument('--run', type=int,
+                        help="Select a single run")
+    parser.add_argument('--start', type=int,
+                        help="Select a starting run")
 
     args = parser.parse_args()
 
@@ -179,7 +182,7 @@ def massive():
                 ('_id', -1))
 
     dt = datetime.timedelta(days=1)
-    t0 = datetime.datetime.utcnow() - 2*dt
+    t0 = datetime.datetime.utcnow() - 2 * dt
 
 
     while True: # yeah yeah
@@ -187,7 +190,7 @@ def massive():
 
         t1 = datetime.datetime.utcnow()
         if t1 - t0 < dt:
-            print("Iterative mode")
+            logging.info("Iterative mode")
 
             # See if there is something to do
             query['start'] = {'$gt' : t0}
@@ -197,6 +200,13 @@ def massive():
             logging.info("Full mode")
             t0 = t1
 
+
+        if args.run:
+            query['number'] = args.run
+
+        if args.start:
+            query['number'] = {'$gte' : args.start}
+
         docs = list(collection.find(query,
                                     sort=sort_key,
                                     projection=['start', 'number','name',
@@ -204,22 +214,23 @@ def massive():
 
         for doc in docs:
 
+            job_name = ''
             if doc['detector'] == 'tpc':
+                job_name = str(doc['number'])
                 job = dict(command='cax --once --run {number} '+config_arg,
-                            number=doc['number'],
-                           )
+                           number=int(job_name))
             elif doc['detector'] == 'muon_veto':
+                job_name = doc['name']
                 job = dict(command='cax --once --name {number} '+config_arg,
-                            number=doc['name'],
-                           )
-         
+                           number=job_name)
+
             script = config.processing_script(job)
 
-            if 'cax_%d_v%s' % (doc['number'], pax.__version__) in qsub.get_queue():
-                logging.debug('Skip: cax_%d_v%s job exists' % (doc['number'], pax.__version__))
+            if 'cax_%s_v%s' % (job_name, pax.__version__) in qsub.get_queue():
+                logging.debug('Skip: cax_%s_v%s job exists' % (job_name, pax.__version__))
                 continue
 
-            while qsub.get_number_in_queue() > (200 if config.get_hostname() == 'midway-login1' else 30):
+            while qsub.get_number_in_queue() > (500 if config.get_hostname() == 'midway-login1' else 30):
                 logging.info("Speed break 60s because %d in queue" % qsub.get_number_in_queue())
                 time.sleep(60)
 
@@ -227,15 +238,15 @@ def massive():
             print(script)
             qsub.submit_job(script)
 
-            logging.debug("Pace by 1s")
+            logging.debug("Pace by 1 s")
             time.sleep(1)  # Pace 1s for batch queue
 
         if run_once:
             break
-        else:
-            pace = 5
-            logging.info("Done, waiting %d minutes" % pace)
-            time.sleep(60*pace) # Pace 5 minutes
+        #else:
+        #    pace = 5
+        #    logging.info("Done, waiting %d minutes" % pace)
+        #    time.sleep(60*pace) # Pace 5 minutes
 
 
 def move():