+++ cax/tasks/process.py
@@ -34,7 +34,7 @@ def _process(name, in_location, host, pax_version, pax_hash,
     print('Welcome to cax-process')
 
     # Import pax so can process the data
-    from pax import core
+    from pax import core, parallel
 
     # Grab the Run DB so we can query it
     collection = config.mongo_collection()
@@ -79,7 +79,9 @@ def _process(name, in_location, host, pax_version, pax_hash,
     # Determine based on run DB what settings to use for processing.
     if doc['detector'] == 'muon_veto':
         pax_config = 'XENON1T_MV'
+        decoder = 'BSON.DecodeZBSON'
     elif doc['detector'] == 'tpc':
+        decoder = 'Pickle.DecodeZPickle'
         if doc['reader']['self_trigger']:
             pax_config = 'XENON1T'
         else:
@@ -88,14 +90,18 @@ def _process(name, in_location, host, pax_version, pax_hash,
     # Try to process data.
     try:
         print('processing', name, in_location, pax_config)
-        p = core.Processor(config_names=pax_config,
-                           config_dict={'pax': {'input_name' : in_location,
-                                                'output_name': output_fullname,
-                                                'n_cpus'     : ncpus
-                                                }
-                                        }
-                           )
-        p.run()
+        pax_kwargs = dict(config_names=pax_config,
+                          config_dict={'pax': {'input_name' : in_location,
+                                               'output_name': output_fullname,
+                                               'decoder_plugin': decoder},
+                                       'DEFAULT': {'lock_breaking_timeout': 600},
+                                       'Queues': {'event_block_size': 1,
+                                                  'max_blocks_on_heap': 1000,
+                                                  'timeout_after_sec': 600}})
+        if ncpus > 1:
+            parallel.multiprocess_locally(n_cpus=ncpus, **pax_kwargs)
+        else:
+            core.Processor(**pax_kwargs).run()
 
     except Exception as exception:
         # Data processing failed.
@@ -128,16 +134,18 @@ class ProcessBatchQueue(Task):
 
     def each_run(self):
         if self.has_tag('donotprocess'):
-            self.log.debug("Do not process tag found")
+            self.log.debug("Do not process tag found, skip processing")
             return
 
         if 'processor' not in self.run_doc or \
                 'DEFAULT' not in self.run_doc['processor']:
+            self.log.debug("processor or DEFAUT tag not in run_doc, skip processing")
             return
 
         processing_parameters = self.run_doc['processor']['DEFAULT']
         if 'gains' not in processing_parameters or \
             'electron_lifetime_liquid' not in processing_parameters:
+            self.log.info("gains or e-lifetime not in run_doc, skip processing")
             return
 
         thishost = config.get_hostname()
@@ -157,10 +165,11 @@ class ProcessBatchQueue(Task):
             return
 
         if self.run_doc['reader']['ini']['write_mode'] != 2:
+            self.log.debug("write_mode != 2, skip processing")
             return
 
         # Get number of events in data set (not set for early runs <1000)
-        events = self.run_doc.get('trigger', {}).get('events_built', -1)
+        events = self.run_doc.get('trigger', {}).get('events_built', 0)
 
         # Skip if 0 events in dataset
         if events == 0:
@@ -216,14 +225,8 @@ class ProcessBatchQueue(Task):
             if 'host' not in datum:
                 continue
 
-            # If the location is Midway SRM...
-            if datum['host']  == "midway-srm":
-                # ... must access from midway-login1
-                if thishost != "midway-login1":
-                    continue
-
-            # Otherwise, if the location doesn't refer to here, skip
-            elif datum['host'] != thishost:
+            # If the location doesn't refer to here, skip
+            if datum['host'] != thishost:
                 continue
 
             # Raw data must exist